+++ app/Containers/User/Configs/user-container.php
@@ -4,16 +4,13 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | User Container
+    | Reset Password
     |--------------------------------------------------------------------------
     |
-    |
+    | Insert your allowed reset password urls to inject into the email.
     |
     */
-
     'allowed-reset-password-urls' => [
-
-        // insert your allowed reset password urls to inject into the email here
         // 'your.domain/resetpassword',
     ],