+++ app/Containers/User/Mails/UserForgotPasswordMail.php
@@ -9,21 +9,35 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 
 /**
  * Class UserForgotPasswordMail
+ *
+ * @author  Sebastian Weckend
  */
 class UserForgotPasswordMail extends Mail implements ShouldQueue
 {
+
     use Queueable;
 
+    /**
+     * @var  \App\Containers\User\Models\User
+     */
     protected $recipient;
+
+    /**
+     * @var string
+     */
     protected $token;
+
+    /**
+     * @var string
+     */
     protected $reseturl;
 
     /**
      * UserForgotPasswordMail constructor.
      *
-     * @param User $recipient
-     * @param $token
-     * @param $reseturl
+     * @param \App\Containers\User\Models\User $recipient
+     * @param                                  $token
+     * @param                                  $reseturl
      */
     public function __construct(User $recipient, $token, $reseturl)
     {
@@ -32,14 +46,17 @@ class UserForgotPasswordMail extends Mail implements ShouldQueue
         $this->reseturl = $reseturl;
     }
 
+    /**
+     * @return  $this
+     */
     public function build()
     {
         return $this->view('user::user-forgotPassword')
             ->to($this->recipient->email, $this->recipient->name)
             ->with([
-                'token' => $this->token,
+                'token'    => $this->token,
                 'reseturl' => $this->reseturl,
-                'email' => $this->recipient->email,
+                'email'    => $this->recipient->email,
             ]);
     }
 }