+++ app/Containers/User/Tasks/CreatePasswordResetTask.php
@@ -3,24 +3,31 @@
 namespace App\Containers\User\Tasks;
 
 use App\Containers\User\Models\User;
-use App\Ship\Exceptions\CreateResourceFailedException;
+use App\Ship\Exceptions\InternalErrorException;
 use App\Ship\Parents\Exceptions\Exception;
 use App\Ship\Parents\Tasks\Task;
 
+/**
+ * Class CreatePasswordResetTask
+ *
+ * @author  Sebastian Weckend
+ */
 class CreatePasswordResetTask extends Task
 {
 
-    public function __construct()
-    {
-        // ..
-    }
-
+    /**
+     * @param \App\Containers\User\Models\User $user
+     *
+     * @return  mixed
+     */
     public function run(User $user)
     {
         try {
-            return app('auth.password.broker')->createToken($user);
+            $token = app('auth.password.broker')->createToken($user);
         } catch (Exception $e) {
-            throw new CreateResourceFailedException;
+            throw (new InternalErrorException())->debug($e);
         }
+
+        return $token;
     }
 }