+++ app/Containers/User/UI/API/Controllers/Controller.php
@@ -5,10 +5,10 @@ namespace App\Containers\User\UI\API\Controllers;
 use Apiato\Core\Foundation\Facades\Apiato;
 use App\Containers\User\UI\API\Requests\CreateAdminRequest;
 use App\Containers\User\UI\API\Requests\DeleteUserRequest;
-use App\Containers\User\UI\API\Requests\ForgotPasswordRequest;
-use App\Containers\User\UI\API\Requests\GetAuthenticatedUserRequest;
 use App\Containers\User\UI\API\Requests\FindUserByIdRequest;
+use App\Containers\User\UI\API\Requests\ForgotPasswordRequest;
 use App\Containers\User\UI\API\Requests\GetAllUsersRequest;
+use App\Containers\User\UI\API\Requests\GetAuthenticatedUserRequest;
 use App\Containers\User\UI\API\Requests\RegisterUserRequest;
 use App\Containers\User\UI\API\Requests\ResetPasswordRequest;
 use App\Containers\User\UI\API\Requests\UpdateUserRequest;
@@ -133,20 +133,24 @@ class Controller extends ApiController
     }
 
     /**
-     * @param ResetPasswordRequest $request
-     * @return \Illuminate\Http\JsonResponse
+     * @param \App\Containers\User\UI\API\Requests\ResetPasswordRequest $request
+     *
+     * @return  \Illuminate\Http\JsonResponse
      */
-    public function resetPassword(ResetPasswordRequest $request) {
+    public function resetPassword(ResetPasswordRequest $request)
+    {
         Apiato::call('User@ResetPasswordAction', [$request]);
 
         return $this->noContent(204);
     }
 
     /**
-     * @param ForgotPasswordRequest $request
-     * @return \Illuminate\Http\JsonResponse
+     * @param \App\Containers\User\UI\API\Requests\ForgotPasswordRequest $request
+     *
+     * @return  \Illuminate\Http\JsonResponse
      */
-    public function forgotPassword(ForgotPasswordRequest $request) {
+    public function forgotPassword(ForgotPasswordRequest $request)
+    {
         Apiato::call('User@ForgotPasswordAction', [$request]);
 
         return $this->noContent(202);