+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added a new config flag (`apiato.requests.force-valid-includes` (default `true`)) to notify users about potential "invalid" `?include` query parameters
 - Added ValueObjects class type to be extended by classes that do not requires to be stored in the DB or have ID.
 - Added a `level` to the roles in order to indicate some kind of hierarchy (e.g., `admin` is "better" than `manager`).
+- Added `/password-forgot` and `/password-reset` endpoints.
 
 ### Changed
 - Changed the `Content-Language` header field (for requesting resources in a specific language) to `Accept-Language` instead (cf. [Specs](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language)).